+++ trionyx/quickstart/app/layouts.py
@@ -1,4 +1,4 @@
 """App layouts"""
-from trionyx.navigation import tabs  # noqa F401
+from trionyx.views import tabs  # noqa F401
 
 # Register and create your tab views